Move benchmark section into its own component
